+++ src/yamldoc/util/misc.py
@@ -23,6 +23,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 # IMPORTS #
 ###########
 
+from typing import Any, Callable, Dict, Tuple
+
 import django.utils.html
 import django.utils.text
 import unidecode
@@ -31,6 +33,10 @@ from django.template.defaultfilters import slugify as default_slugify
 #######################
 # INTERFACE FUNCTIONS #
 #######################
+
+
+# Raw represents data for a Django model, stored in serialized text format.
+Raw = Dict[str, Any]
 
 
 def slugify(string):
@@ -48,3 +54,58 @@ def slugify(string):
         raise ValueError(s.format(string, clean))
 
     return slug
+
+
+def copy_in_order(a: Raw, b: Raw) -> Raw:
+    """Copy all entries in a to b in their existing order. Return b.
+
+    It is assumed here that b already contains some of the entries from a in a
+    significant order, but not necessarily all, and/or some entries not in a.
+
+    """
+    for k in a:
+        b[k] = a[k]
+    return b
+
+
+def field_order_fn(fields: Tuple[str],
+                   finalizer: Callable[[Raw, Raw], Raw] = copy_in_order):
+    """Close over a mapping function for yamlwrap.
+
+    “fields” is expected to be tuple(f.name for f in model._meta.fields) or an
+    equivalent tuple of strings naming significant database fields in a
+    preferred order for storage in YAML.
+
+    With the default “finalizer”, preserve even those entries that are not
+    named in “fields”, but put them after “fields”, without changing their
+    internal order.
+
+    """
+    def order(fragment: Raw) -> Raw:
+        """Produce a dictionary for replacement of another.
+
+        Any metadata or raw data not named like database fields is retained,
+        with its internal order unchanged.
+
+        """
+        ordered: Raw = {}
+        for f in fields:
+            try:
+                ordered[f] = fragment[f]
+            except KeyError:
+                pass
+        return finalizer(fragment, ordered)
+    return order
+
+
+def unique_alphabetizer(key: str):
+    """Close over a mapping function for yamlwrap."""
+    def order(fragment: Raw):
+        """Sort raw data for a field alphabetically and remove duplicates."""
+        if key in fragment:
+            fragment[key] = sorted(set(fragment[key]))
+        return fragment
+    return order
+
+
+alphabetize_tags = unique_alphabetizer('tags')  # Example.